% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{addslash_ifneeded}
\alias{addslash_ifneeded}
\title{addslash_ifneeded}
\usage{
addslash_ifneeded(folder)
}
\arguments{
\item{folder}{folder where is located the hdf5 file.}
}
\description{
This function allows you add slash symbol, if it is needed, to the path of folder which contains the hdf5 file.
}
\examples{
addslash_ifneeded(folder="synth_cmd/synth/")
}
\keyword{folder}
