% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{hdf5_file_from_pars}
\alias{hdf5_file_from_pars}
\title{hdf5_file_from_pars}
\usage{
hdf5_file_from_pars(pars, folder)
}
\arguments{
\item{folder}{folder where is located the hdf5 file.}
}
\description{
This function allows you create a path to the .hdf5 file in the folder given the parameters in pars variable. You have to indicate parameters and folder.
}
\examples{
hdf5_file_from_pars(c(1,23,2),folder="synth_cmd/synth/")
}
\keyword{folder}
