% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_name_from_pars.R,
%   R/montecarlizza_packages.R
\name{hdf5_name_from_pars}
\alias{hdf5_name_from_pars}
\title{hdf5_name_from_pars}
\usage{
hdf5_name_from_pars(pars)

hdf5_name_from_pars(pars)
}
\arguments{
\item{pars}{Cluster parameters Age BinaryFraction SigmaHB.}
}
\description{
This function allows you create the file name from parameters.

This function allows you create the file name from parameters.
}
\examples{
hdf5_name_from_pars()
hdf5_name_from_pars(c(1,23,2))
}
\keyword{parameters}
\keyword{parss}
