% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{list_hdf5_file}
\alias{list_hdf5_file}
\title{list_hdf5_file}
\usage{
list_hdf5_file(folder)
}
\arguments{
\item{folder}{folder where are located the hdf5 files.}
}
\description{
This function allows you get list of hdf5 file in the folder.
}
\examples{
list_hdf5_file(folder="synth_cmd/synth")
}
\keyword{files.}
\keyword{folder,}
\keyword{hdf5}
