% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{pars_from_hdf5_file}
\alias{pars_from_hdf5_file}
\title{pars_from_hdf5_file}
\usage{
pars_from_hdf5_file(name)
}
\arguments{
\item{pars}{Cluster parameters Age BinaryFraction SigmaHB.}
}
\description{
This function allows you get the cluster parameters from the hdf5 file name.
}
\examples{
pars_from_hdf5_file("3.33_45_0.02.hdf5")
}
\keyword{pars}
