% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{pars_to_name_diccionary}
\alias{pars_to_name_diccionary}
\title{pars_to_name_diccionary}
\usage{
pars_to_name_diccionary(folder)
}
\arguments{
\item{folder}{folder where are located the hdf5 files.}
}
\description{
This function allows you create a diccionary of parameters.
}
\examples{
pars_to_name_diccionary(folder="synth_cmd/synth")
}
\keyword{files.}
\keyword{folder,}
\keyword{hdf5}
