% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlizza_packages.R
\name{read_from_hdf5}
\alias{read_from_hdf5}
\title{read_from_hdf5}
\usage{
read_from_hdf5(pars)
}
\arguments{
\item{pars}{Cluster parameters Age BinaryFraction SigmaHB.}
}
\description{
This function allows you read the hdf5 file needed.
}
\examples{
read_from_hdf5(c(1,23,2))
}
\keyword{pars}
